// src/pages/english/enums.ts
export enum ESentenceTypes {
  QUESTION = 'question',
  STATEMENT = 'statement',
  NEGATION = 'negation',
}

export enum ESentenceTypesRussian {
  QUESTION = 'вопрос',
  STATEMENT = 'утверждение',
  NEGATION = 'отрицание',
}

export enum ETimes {
  FUTURE = 'future',
  PRESENT = 'present',
  PAST = 'past',
}

export enum ETimesRussian {
  FUTURE = 'будущее время',
  PRESENT = 'настоящее время',
  PAST = 'прошедшее время',
}

export enum EPronouns {
  I = 'I',
  YOU = 'you',
  WE = 'we',
  THEY = 'they',
  HE = 'he',
  SHE = 'she',
}

export enum EResult {
  CORRECT = 'correct',
  WRONG = 'wrong',
  NONE = 'none',
}

export enum EVerbForms {
  REGULAR = 'regular',
  WRONG = 'wrong',
}